changed course on quickAction and added date-based update notifications
Whenever I push an update, I can set an expiration date. Until that date, the RQ Mods link will be highlighted, encouraging users to click it.

For the custom quick action, I'm just going to add that as an option for the custom quick links. There's no room to put the action button anyway.

Added an "update checklist" to be followed every time an update is pushed to the chrome store.